refactor: 클래스 명 변경 WinningLottery -> WinningLotteryRank WinningLotteryNumbers -> WinningLottery
